Extracted 'na' logic into a separate module
